extract common model saving&conversion fn
